+++ CHANGES.md
@@ -36,6 +36,7 @@
 * Runtime: fix fake filesystem with path containing special regexp chars. (#2132)
 * Runtime/wasm: fix unmarshalling of compressed data (#2141)
 * Runtime: fix compilation of loops at start of exception handlers (#2151)
+* Compiler: fix parallel renaming (#2156)
 
 # 6.2.0 (2025-07-30) - Lille